move SDK internals into core and discovery namespaces
- add blesta_sdk.core.{client,async_client,response,errors,config,
  pagination,redaction,validation,dateutil,retry} with canonical code
- add blesta_sdk.discovery.registry with BlestaDiscovery and MethodSpec
- add blesta_sdk.core.__init__ and blesta_sdk.discovery.__init__
  re-exporting key classes
- replace _*.py private modules with compatibility shims that
  re-export from new namespaces; preserve all patchable module-level
  names (asyncio, time, json, random) for existing test patches
- update __init__.py to import from new canonical paths
- update test patches: blesta_sdk._discovery -> blesta_sdk.discovery.registry,
  sys.modules key for httpx simulation test
- add TestCoreNamespaceImports covering old and new import paths
- 827 tests pass, 98.73% coverage
