+++ src/blesta_sdk/core/__init__.py
@@ -1,0 +1,29 @@
+"""Core HTTP client, response parsing, and shared helpers for blesta_sdk."""
+
+from __future__ import annotations
+
+from blesta_sdk.core.client import BlestaRequest
+from blesta_sdk.core.config import BlestaEnvConfig
+from blesta_sdk.core.errors import (
+    BlestaAPIError,
+    BlestaAuthError,
+    BlestaConnectionError,
+    BlestaError,
+    BlestaRateLimitError,
+    BlestaServerError,
+    PaginationError,
+)
+from blesta_sdk.core.response import BlestaResponse
+
+__all__ = [
+    "BlestaAPIError",
+    "BlestaAuthError",
+    "BlestaConnectionError",
+    "BlestaEnvConfig",
+    "BlestaError",
+    "BlestaRateLimitError",
+    "BlestaRequest",
+    "BlestaResponse",
+    "BlestaServerError",
+    "PaginationError",
+]
