+++ src/blesta_sdk/core/client.py
@@ -1,0 +1,727 @@
+"""Sync HTTP client for the Blesta REST API.
+
+This module is not part of the public API. Import
+:class:`~blesta_sdk.BlestaRequest` from ``blesta_sdk`` directly.
+"""
+
+from __future__ import annotations
+
+import logging
+import time
+from collections.abc import Iterator
+from typing import Any, Literal
+
+import requests
+from requests.adapters import HTTPAdapter
+from requests.auth import HTTPBasicAuth
+
+from blesta_sdk.core.dateutil import _month_boundaries
+from blesta_sdk.core.pagination import PaginationState
+from blesta_sdk.core.redaction import redact_args
+from blesta_sdk.core.response import BlestaResponse
+from blesta_sdk.core.retry import jitter_delay
+from blesta_sdk.core.validation import validate_segment
+
+logger = logging.getLogger(__name__)
+
+DEFAULT_TIMEOUT = 30
+
+_IDEMPOTENT_METHODS = frozenset({"GET", "DELETE"})
+
+
+class BlestaRequest:
+    """HTTP client for the Blesta REST API.
+
+    Wraps :class:`requests.Session` with Basic Auth and provides
+    convenience methods for GET/POST/PUT/DELETE, automatic pagination,
+    and report fetching.
+
+    Can be used as a context manager::
+
+        with BlestaRequest(url, user, key) as api:
+            response = api.get("clients", "getList")
+
+    :param url: Base URL of the Blesta API (e.g., ``"https://example.com/api"``).
+    :param user: API username.
+    :param key: API key.
+    :param timeout: Request timeout in seconds.
+    :param max_retries: Number of retries on transient failures.
+        GET and DELETE retry on network errors, 5xx responses, and 429.
+        POST and PUT retry **only on 429** (rate-limit) — never on 5xx,
+        because a server error does not guarantee the write failed and
+        retrying risks duplicate billing records.
+        Uses exponential backoff with jitter. Defaults to ``0`` (no retries).
+    :param retry_mutations: Include POST and PUT in the retry loop.
+        When ``True``, POST/PUT will retry on 429 but still never on 5xx.
+        Defaults to ``False``.
+    :param pool_connections: Number of connection pools to cache.
+        Defaults to ``10``.
+    :param pool_maxsize: Maximum number of connections per pool.
+        Defaults to ``10``.
+    :param auth_method: Authentication method. ``"basic"`` uses HTTP Basic
+        Auth. ``"header"`` sends credentials via ``BLESTA-API-USER`` and
+        ``BLESTA-API-KEY`` headers (recommended by Blesta, requires no
+        server-side CGI/PHP-FPM configuration). Defaults to ``"basic"``.
+    :param raise_on_error: When ``True``, :meth:`submit` calls
+        :meth:`~blesta_sdk.BlestaResponse.raise_for_status` before
+        returning, raising a :class:`~blesta_sdk.BlestaError` subclass
+        on non-success responses. Defaults to ``False``.
+    :param allow_http: When ``True``, permits ``http://`` base URLs.
+        Defaults to ``False``. HTTP sends credentials in plaintext —
+        only enable this for local development or explicit test
+        environments.
+    :param discovery: Optional :class:`~blesta_sdk.BlestaDiscovery` instance
+        to use instead of the module-level singleton. Useful when loading
+        schemas from a custom path or when injecting a mock in tests.
+    """
+
+    def __init__(
+        self,
+        url: str,
+        user: str,
+        key: str,
+        timeout: int | float = DEFAULT_TIMEOUT,
+        max_retries: int = 0,
+        retry_mutations: bool = False,
+        pool_connections: int = 10,
+        pool_maxsize: int = 10,
+        auth_method: Literal["basic", "header"] = "basic",
+        raise_on_error: bool = False,
+        allow_http: bool = False,
+        discovery: Any = None,
+    ):
+        if url.startswith("http://") and not allow_http:
+            raise ValueError(
+                "base_url uses HTTP which sends credentials in plaintext. "
+                "Pass allow_http=True to explicitly permit this (local/dev only)."
+            )
+        self.base_url = url.rstrip("/") + "/"
+        self.user = user
+        self.key = key
+        self.timeout = timeout
+        self.max_retries = max_retries
+        self.retry_mutations = retry_mutations
+        self.auth_method = auth_method
+        self.raise_on_error = raise_on_error
+        self._discovery = discovery
+        self._last_request: dict[str, Any] | None = None
+        self.session = requests.Session()
+        if auth_method == "header":
+            self.session.headers["BLESTA-API-USER"] = self.user
+            self.session.headers["BLESTA-API-KEY"] = self.key
+        else:
+            self.session.auth = HTTPBasicAuth(self.user, self.key)
+        adapter = HTTPAdapter(
+            pool_connections=pool_connections,
+            pool_maxsize=pool_maxsize,
+        )
+        self.session.mount("https://", adapter)
+        self.session.mount("http://", adapter)
+
+    def __enter__(self) -> BlestaRequest:
+        return self
+
+    def __exit__(self, *args: Any) -> None:
+        self.session.close()
+
+    def __repr__(self) -> str:
+        return f"BlestaRequest(url={self.base_url!r}, user={self.user!r})"
+
+    def close(self) -> None:
+        """Close the underlying HTTP session."""
+        self.session.close()
+
+    @staticmethod
+    def _validate_segment(segment: str, name: str) -> None:
+        validate_segment(segment, name)
+
+    def _get_discovery(self) -> Any:
+        """Return the BlestaDiscovery instance for this client.
+
+        Returns the instance passed at construction time if provided,
+        otherwise falls back to the module-level singleton.
+        """
+        if self._discovery is not None:
+            return self._discovery
+        from blesta_sdk.discovery.registry import _get_discovery
+
+        return _get_discovery()
+
+    def get(
+        self, model: str, method: str, args: dict[str, Any] | None = None
+    ) -> BlestaResponse:
+        """Send a GET request. Parameters are passed as query strings.
+
+        :param model: API model (e.g., ``"clients"``).
+        :param method: API method (e.g., ``"getList"``).
+        :param args: Query parameters.
+        :return: Parsed API response.
+        """
+        return self.submit(model, method, args, "GET")
+
+    def post(
+        self, model: str, method: str, args: dict[str, Any] | None = None
+    ) -> BlestaResponse:
+        """Send a POST request. Parameters are sent as a JSON body.
+
+        :param model: API model (e.g., ``"clients"``).
+        :param method: API method (e.g., ``"create"``).
+        :param args: JSON body data.
+        :return: Parsed API response.
+        """
+        return self.submit(model, method, args, "POST")
+
+    def put(
+        self, model: str, method: str, args: dict[str, Any] | None = None
+    ) -> BlestaResponse:
+        """Send a PUT request. Parameters are sent as a JSON body.
+
+        :param model: API model (e.g., ``"clients"``).
+        :param method: API method (e.g., ``"edit"``).
+        :param args: JSON body data.
+        :return: Parsed API response.
+        """
+        return self.submit(model, method, args, "PUT")
+
+    def delete(
+        self, model: str, method: str, args: dict[str, Any] | None = None
+    ) -> BlestaResponse:
+        """Send a DELETE request. Parameters are sent as a JSON body.
+
+        :param model: API model (e.g., ``"clients"``).
+        :param method: API method (e.g., ``"delete"``).
+        :param args: JSON body data.
+        :return: Parsed API response.
+        """
+        return self.submit(model, method, args, "DELETE")
+
+    def submit(
+        self,
+        model: str,
+        method: str,
+        args: dict[str, Any] | None = None,
+        action: Literal["GET", "POST", "PUT", "DELETE"] = "POST",
+    ) -> BlestaResponse:
+        """Send an HTTP request to the Blesta API.
+
+        Prefer :meth:`get`, :meth:`post`, :meth:`put`, or :meth:`delete`
+        over calling this method directly.
+
+        On network errors, returns a ``BlestaResponse`` with
+        ``status_code=0`` and the exception message as raw text.
+
+        Retries are only attempted for idempotent methods (GET, DELETE)
+        unless ``retry_mutations=True`` was passed to the constructor.
+
+        :param model: API model (e.g., ``"clients"``). For plugin models
+            use dot notation: ``"plugin.model"`` (builds
+            ``plugin.model/method.json``).
+        :param method: API method (e.g., ``"getList"``).
+        :param args: Query parameters (GET) or JSON body (POST/PUT/DELETE).
+        :param action: HTTP method — ``"GET"``, ``"POST"``, ``"PUT"``, or ``"DELETE"``.
+        :return: Parsed API response.
+        :raises ValueError: If *action* is not a recognized HTTP method.
+        """
+        if args is None:
+            args = {}
+
+        self._validate_segment(model, "model")
+        self._validate_segment(method, "method")
+        url = f"{self.base_url}{model}/{method}.json"
+        self._last_request = {"url": url, "args": args.copy()}
+
+        can_retry = self.retry_mutations or action in _IDEMPOTENT_METHODS
+        effective_retries = self.max_retries if can_retry else 0
+
+        last_response: BlestaResponse | None = None
+        for attempt in range(effective_retries + 1):
+            try:
+                if action == "GET":
+                    response = self.session.get(url, params=args, timeout=self.timeout)
+                elif action == "POST":
+                    response = self.session.post(url, json=args, timeout=self.timeout)
+                elif action == "PUT":
+                    response = self.session.put(url, json=args, timeout=self.timeout)
+                elif action == "DELETE":
+                    response = self.session.delete(url, json=args, timeout=self.timeout)
+                else:
+                    raise ValueError("Invalid HTTP action specified.")
+
+                last_response = BlestaResponse(
+                    response.text, response.status_code, response.headers
+                )
+
+                # Mutations (POST/PUT) must never be retried on 5xx — a server
+                # error does not guarantee the write failed, and retrying risks
+                # duplicate billing records. Only 429 (rate-limit) is safe to
+                # retry for mutations because the request was explicitly rejected
+                # before being processed.
+                is_mutation = action in ("POST", "PUT")
+                if is_mutation:
+                    is_retriable = response.status_code == 429
+                else:
+                    is_retriable = (
+                        response.status_code >= 500 or response.status_code == 429
+                    )
+                if not is_retriable or attempt == effective_retries:
+                    if self.raise_on_error:
+                        last_response.raise_for_status()
+                    return last_response
+
+                logger.warning(
+                    "Retry %d/%d: HTTP %d from %s",
+                    attempt + 1,
+                    effective_retries,
+                    response.status_code,
+                    url,
+                )
+
+            except requests.RequestException as e:
+                logger.error("Request failed: %s", e)
+                last_response = BlestaResponse(str(e), 0)
+
+                if attempt == effective_retries:
+                    if self.raise_on_error:
+                        last_response.raise_for_status()
+                    return last_response
+
+                logger.warning("Retry %d/%d: %s", attempt + 1, effective_retries, e)
+
+            if last_response is not None and last_response.status_code == 429:
+                try:
+                    retry_after = int(last_response.headers.get("Retry-After", ""))
+                except (ValueError, TypeError):
+                    retry_after = 0
+                if retry_after > 0:
+                    time.sleep(retry_after)
+                    continue
+
+            time.sleep(jitter_delay(attempt))
+
+        if last_response is None:  # pragma: no cover
+            raise RuntimeError("Retry loop exited without a response")
+        return last_response  # pragma: no cover
+
+    def iter_all(
+        self,
+        model: str,
+        method: str,
+        args: dict[str, Any] | None = None,
+        start_page: int = 1,
+        max_pages: int | None = None,
+        on_error: Literal["raise", "warn"] = "warn",
+    ) -> Iterator[Any]:
+        """Yield individual items across all pages.
+
+        Calls the API with ``page=1``, ``page=2``, etc. until an empty
+        response is returned.
+
+        :param model: API model (e.g., ``"invoices"``).
+        :param method: API method (e.g., ``"getList"``).
+        :param args: Query parameters (``page`` is managed automatically).
+        :param start_page: Page number to start from.
+        :param max_pages: Maximum number of pages to fetch. ``None``
+            means no limit.
+        :param on_error: Behavior on non-200 status codes. ``"raise"``
+            raises :class:`~blesta_sdk.PaginationError` with partial
+            results attached. ``"warn"`` logs a warning and stops
+            iteration (backward-compatible default).
+        :return: Iterator of individual result items.
+        :raises PaginationError: If *on_error* is ``"raise"`` and a
+            non-200 response is received.
+        """
+        base_args = args or {}
+        state = PaginationState(start_page, max_pages, on_error)
+
+        while state.has_next_page():
+            response = self.get(model, method, {**base_args, "page": state.page})
+            if state.check_response(response):
+                return
+            data = response.data
+            if state.check_data(data):
+                return
+            state.collect(data)
+            if isinstance(data, list):
+                yield from data
+            else:
+                yield data
+                return
+            state.advance()
+
+    def get_all(
+        self,
+        model: str,
+        method: str,
+        args: dict[str, Any] | None = None,
+        start_page: int = 1,
+        max_pages: int | None = None,
+        on_error: Literal["raise", "warn"] = "warn",
+    ) -> list[Any]:
+        """Fetch all pages and return results as a single list.
+
+        Convenience wrapper around :meth:`iter_all`.
+
+        .. warning::
+            Materializes all records into memory. For large datasets
+            (100k+ records), prefer :meth:`iter_all` or
+            :meth:`iter_pages` to process records in a streaming
+            fashion.
+
+        :param model: API model (e.g., ``"invoices"``).
+        :param method: API method (e.g., ``"getList"``).
+        :param args: Query parameters (``page`` is managed automatically).
+        :param start_page: Page number to start from.
+        :param max_pages: Maximum number of pages to fetch. ``None``
+            means no limit.
+        :param on_error: Behavior on non-200 status codes. ``"raise"``
+            raises :class:`~blesta_sdk.PaginationError` with partial
+            results attached. ``"warn"`` logs a warning and stops
+            iteration (backward-compatible default).
+        :return: List of all result items across all pages.
+        :raises PaginationError: If *on_error* is ``"raise"`` and a
+            non-200 response is received.
+        """
+        return list(self.iter_all(model, method, args, start_page, max_pages, on_error))
+
+    def iter_pages(
+        self,
+        model: str,
+        method: str,
+        args: dict[str, Any] | None = None,
+        start_page: int = 1,
+        max_pages: int | None = None,
+        on_error: Literal["raise", "warn"] = "warn",
+    ) -> Iterator[list[Any]]:
+        """Yield each page of results as a separate list.
+
+        Unlike :meth:`iter_all` (which yields individual items),
+        this method yields one list per API page — useful for
+        batch-flushing to a database or file.
+
+        :param model: API model (e.g., ``"invoices"``).
+        :param method: API method (e.g., ``"getList"``).
+        :param args: Query parameters (``page`` is managed automatically).
+        :param start_page: Page number to start from.
+        :param max_pages: Maximum number of pages to fetch.
+        :param on_error: Behavior on non-200 status codes. ``"raise"``
+            raises :class:`~blesta_sdk.PaginationError` with partial
+            page count. ``"warn"`` logs a warning and stops
+            iteration (backward-compatible default).
+        :return: Iterator of page lists.
+        :raises PaginationError: If *on_error* is ``"raise"`` and a
+            non-200 response is received.
+        """
+        base_args = args or {}
+        state = PaginationState(start_page, max_pages, on_error)
+
+        while state.has_next_page():
+            response = self.get(model, method, {**base_args, "page": state.page})
+            if state.check_response(response):
+                return
+            data = response.data
+            if state.check_data(data):
+                return
+            state.collect(data)
+            if isinstance(data, list):
+                yield data
+            else:
+                yield [data]
+                return
+            state.advance()
+
+    def count(
+        self,
+        model: str,
+        method: str = "getListCount",
+        args: dict[str, Any] | None = None,
+    ) -> int:
+        """Fetch a record count from a Blesta ``*Count`` method.
+
+        Many Blesta models expose a ``getListCount`` method that returns
+        a single integer.  This method wraps that call and returns the
+        count as a Python :class:`int`.
+
+        :param model: API model (e.g., ``"transactions"``).
+        :param method: Count method name.  Defaults to ``"getListCount"``.
+        :param args: Query parameters.
+        :return: Record count, or ``0`` on errors.
+        """
+        response = self.get(model, method, args)
+        if response.status_code != 200:
+            logger.warning(
+                "count() got HTTP %d for %s/%s", response.status_code, model, method
+            )
+            return 0
+        data = response.data
+        if data is None:
+            return 0
+        try:
+            return int(data)
+        except (TypeError, ValueError):
+            logger.warning(
+                "count() expected int-like response from %s/%s, got %s: %r",
+                model,
+                method,
+                type(data).__name__,
+                data,
+            )
+            return 0
+
+    def extract(
+        self,
+        targets: list[tuple[str, str] | tuple[str, str, dict[str, Any]]],
+    ) -> dict[str, list[Any]]:
+        """Fetch multiple paginated endpoints and return results keyed by model.
+
+        Convenience method for ETL workflows that pull several models at once.
+        Each target is a tuple of ``(model, method)`` or
+        ``(model, method, args)``.
+
+        :param targets: List of extraction targets.
+        :return: Dict mapping ``"model.method"`` to list of results.
+        """
+        results: dict[str, list[Any]] = {}
+        for target in targets:
+            if len(target) == 3:
+                model, method, args = target  # type: ignore[misc]
+            else:
+                model, method = target  # type: ignore[misc]
+                args = None
+            key = f"{model}.{method}"
+            results[key] = self.get_all(model, method, args)
+        return results
+
+    def get_report(
+        self,
+        report_type: str,
+        start_date: str,
+        end_date: str,
+        extra_vars: dict[str, str] | None = None,
+    ) -> BlestaResponse:
+        """Fetch a Blesta report via ``report_manager/fetchAll``.
+
+        Automatically formats the ``vars[]`` parameters that Blesta
+        expects. Report responses are typically CSV — use
+        :attr:`~BlestaResponse.csv_data` to parse them.
+
+        :param report_type: Report type (e.g., ``"package_revenue"``).
+        :param start_date: Start date as ``"YYYY-MM-DD"``.
+        :param end_date: End date as ``"YYYY-MM-DD"``.
+        :param extra_vars: Additional ``vars[]`` parameters. Keys are
+            auto-wrapped in ``vars[...]`` unless already wrapped.
+        :return: Parsed API response (usually CSV).
+        """
+        args: dict[str, str] = {
+            "type": report_type,
+            "vars[start_date]": start_date,
+            "vars[end_date]": end_date,
+        }
+        if extra_vars:
+            for key, value in extra_vars.items():
+                param_key = key if key.startswith("vars[") else f"vars[{key}]"
+                args[param_key] = value
+
+        return self.get("report_manager", "fetchAll", args)
+
+    def get_report_series_pages(
+        self,
+        report_type: str,
+        start_month: str,
+        end_month: str,
+        extra_vars: dict[str, str] | None = None,
+    ) -> Iterator[tuple[str, BlestaResponse]]:
+        """Yield ``(period, response)`` for each month in a date range.
+
+        Fetches one report per month via :meth:`get_report`. Yields
+        all months including those that return errors, so the caller
+        can decide how to handle failures.
+
+        :param report_type: Report type (e.g., ``"package_revenue"``).
+        :param start_month: Start month as ``"YYYY-MM"`` (inclusive).
+        :param end_month: End month as ``"YYYY-MM"`` (inclusive).
+        :param extra_vars: Additional ``vars[]`` parameters.
+        :return: Iterator of ``(period, BlestaResponse)`` tuples.
+        :raises ValueError: If *start_month* is after *end_month* or
+            the format is invalid.
+        """
+        boundaries = _month_boundaries(start_month, end_month)
+        for first_day, last_day, period in boundaries:
+            logger.debug("Fetching report %r for %s", report_type, period)
+            response = self.get_report(report_type, first_day, last_day, extra_vars)
+            yield (period, response)
+
+    def get_report_series(
+        self,
+        report_type: str,
+        start_month: str,
+        end_month: str,
+        extra_vars: dict[str, str] | None = None,
+    ) -> list[dict[str, str]]:
+        """Fetch monthly reports and return all rows as a flat list.
+
+        Convenience wrapper around :meth:`get_report_series_pages`.
+        Each returned row dict has a ``"_period"`` key added with the
+        ``"YYYY-MM"`` value. Months that return errors or non-CSV
+        responses are skipped with a warning log.
+
+        :param report_type: Report type (e.g., ``"package_revenue"``).
+        :param start_month: Start month as ``"YYYY-MM"`` (inclusive).
+        :param end_month: End month as ``"YYYY-MM"`` (inclusive).
+        :param extra_vars: Additional ``vars[]`` parameters.
+        :return: Flat list of row dicts from all months.
+        :raises ValueError: If *start_month* is after *end_month* or
+            the format is invalid.
+        """
+        rows: list[dict[str, str]] = []
+        for period, response in self.get_report_series_pages(
+            report_type, start_month, end_month, extra_vars
+        ):
+            if response.status_code != 200:
+                logger.warning(
+                    "Report %r for %s: HTTP %d, skipping",
+                    report_type,
+                    period,
+                    response.status_code,
+                )
+                continue
+            csv_rows = response.csv_data
+            if csv_rows is None:
+                logger.warning(
+                    "Report %r for %s: no CSV data in response, skipping",
+                    report_type,
+                    period,
+                )
+                continue
+            for row in csv_rows:
+                rows.append({**row, "_period": period})
+        return rows
+
+    def call(
+        self,
+        model: str,
+        method: str,
+        args: dict[str, Any] | None = None,
+        action: str | None = None,
+    ) -> BlestaResponse:
+        """Call an API method, inferring the HTTP method from the schema.
+
+        Uses :class:`~blesta_sdk.BlestaDiscovery` to resolve the correct
+        HTTP method when *action* is ``None``. If the schema cannot
+        resolve the method, falls back to a safe prefix-based heuristic
+        (e.g. ``get*`` -> GET, ``create*`` -> POST, ``delete*`` -> DELETE).
+        If the HTTP method cannot be determined from either source,
+        :exc:`ValueError` is raised — pass *action* explicitly to avoid this.
+
+        :param model: API model (e.g., ``"clients"``).
+        :param method: API method (e.g., ``"getList"``).
+        :param args: Request parameters.
+        :param action: Explicit HTTP method override. If ``None``,
+            the method is inferred from the schema or method name prefix.
+        :return: Parsed API response.
+        :raises ValueError: When the HTTP method cannot be determined from
+            the schema or method name prefix and *action* is ``None``.
+        """
+        if action is None:
+            from blesta_sdk.discovery.registry import _infer_http_method
+
+            _sentinel = "_UNRESOLVED_"
+            disco = self._get_discovery()
+            action = disco.resolve_http_method(model, method, default=_sentinel)
+            if action == _sentinel:
+                inferred = _infer_http_method(method)
+                if inferred is not None:
+                    action = inferred
+                else:
+                    raise ValueError(
+                        f"call({model!r}, {method!r}): HTTP method could not be"
+                        " determined from schema or method name prefix. Pass"
+                        " action= explicitly."
+                    )
+        return self.submit(model, method, args, action)  # type: ignore[arg-type]
+
+    def call_all(
+        self,
+        model: str,
+        method: str,
+        args: dict[str, Any] | None = None,
+        start_page: int = 1,
+    ) -> list[Any]:
+        """Paginate an API method, validating via schema that it uses GET.
+
+        Convenience wrapper around :meth:`get_all` that uses
+        :class:`~blesta_sdk.BlestaDiscovery` to confirm the method resolves
+        to GET before paginating. If the schema is unavailable, proceeds
+        with a debug log note. If the schema definitively says the method
+        is not GET (e.g., POST, PUT, DELETE), :exc:`ValueError` is raised.
+
+        :param model: API model (e.g., ``"invoices"``).
+        :param method: API method (e.g., ``"getList"``).
+        :param args: Query parameters.
+        :param start_page: Page number to start from.
+        :return: List of all result items across all pages.
+        :raises ValueError: When the schema indicates the method is not GET.
+        """
+        _sentinel = "_UNRESOLVED_"
+        disco = self._get_discovery()
+        http_method = disco.resolve_http_method(model, method, default=_sentinel)
+        if http_method == _sentinel:
+            logger.debug(
+                "call_all(%s, %s): schema unavailable, proceeding with GET.",
+                model,
+                method,
+            )
+        elif http_method != "GET":
+            raise ValueError(
+                f"call_all({model!r}, {method!r}): schema says HTTP method is"
+                f" {http_method!r}, not GET. Use get_all() directly or pass the"
+                " correct pagination method."
+            )
+        return self.get_all(model, method, args, start_page)
+
+    def count_for(
+        self,
+        model: str,
+        list_method: str = "getList",
+        args: dict[str, Any] | None = None,
+    ) -> int:
+        """Fetch the record count for a paginated list method.
+
+        Uses :class:`~blesta_sdk.BlestaDiscovery` to find the matching
+        count method (e.g., ``"getList"`` -> ``"getListCount"``). Falls
+        back to ``list_method + "Count"`` if the schema is unavailable,
+        with a warning log so callers can verify the endpoint exists.
+
+        :param model: API model (e.g., ``"transactions"``).
+        :param list_method: The list method to find a count for.
+        :param args: Query parameters.
+        :return: Record count, or ``0`` on errors.
+        """
+        disco = self._get_discovery()
+        count_method = disco.suggest_pagination_pair(model, list_method)
+        if count_method is None:
+            count_method = list_method + "Count"
+            logger.warning(
+                "count_for(%s, %s): no count method found in schema, "
+                "falling back to %r. Verify this endpoint exists.",
+                model,
+                list_method,
+                count_method,
+            )
+        return self.count(model, count_method, args)
+
+    def get_last_request(self) -> dict[str, Any] | None:
+        """Return details of the last request made.
+
+        The ``"args"`` value has sensitive keys redacted (replaced with
+        ``"***"``) to prevent accidental credential leakage in logs or CLI
+        output. The actual request payload is not affected.
+
+        :return: Dict with ``"url"`` and ``"args"`` keys, or ``None``
+            if no requests have been made.
+        """
+        if self._last_request is None:
+            return None
+        return {
+            "url": self._last_request["url"],
+            "args": redact_args(self._last_request["args"]),
+        }
