+++ src/blesta_sdk/core/validation.py
@@ -1,0 +1,41 @@
+"""Input validation helpers for URL path segments."""
+
+from __future__ import annotations
+
+
+def validate_segment(segment: str, name: str) -> None:
+    """Validate a URL path segment (model or method name).
+
+    Rejects segments that could escape the API path via scheme
+    injection, path traversal, netloc override, or backslash tricks.
+    Allows plugin dot notation (e.g. ``"plugin.model"``).
+
+    :param segment: The path segment to validate.
+    :param name: Human-readable label for error messages (e.g. ``"model"``).
+    :raises ValueError: If the segment is unsafe.
+    """
+    if not segment:
+        raise ValueError(f"{name} cannot be empty")
+    if "\x00" in segment:
+        raise ValueError(f"{name} cannot contain null bytes")
+    if any(c.isspace() for c in segment):
+        raise ValueError(f"{name} cannot contain whitespace")
+    if "://" in segment:
+        raise ValueError(f"{name} cannot contain scheme")
+    if segment.startswith("//"):
+        raise ValueError(f"{name} cannot start with '//'")
+    if segment.startswith("/"):
+        raise ValueError(f"{name} cannot start with '/'")
+    if ".." in segment:
+        raise ValueError(f"{name} cannot contain '..'")
+    if "\\" in segment:
+        raise ValueError(f"{name} cannot contain backslashes")
+    if "?" in segment:
+        raise ValueError(f"{name} cannot contain '?'")
+    if "#" in segment:
+        raise ValueError(f"{name} cannot contain '#'")
+    if "%" in segment:
+        raise ValueError(
+            f"{name} cannot contain percent-encoded characters "
+            "(e.g. %2F, %2e%2e, %00)"
+        )
