+++ src/blesta_sdk/discovery/__init__.py
@@ -1,0 +1,10 @@
+"""Schema-driven API discovery for the Blesta SDK."""
+
+from __future__ import annotations
+
+from blesta_sdk.discovery.registry import BlestaDiscovery, MethodSpec
+
+__all__ = [
+    "BlestaDiscovery",
+    "MethodSpec",
+]
